Fixed modules' structure to pass tests
