--- homework1/task03/test_file.txt
@@ -1,11 +1,0 @@
-1
-423
-5
-3
-9
-0
--5634
-7
--9
-2
-436
