+++ homework1/task04/__init__.py
@@ -1,0 +1,19 @@
+"""
+Classic task, a kind of walnut for you
+
+Given four lists A, B, C, D of integer values,
+    compute how many tuples (i, j, k, l) there are such that A[i] + B[j] + C[k] + D[l] is zero.
+
+We guarantee, that all A, B, C, D have same length of N where 0 ≤ N ≤ 1000.
+"""
+
+from itertools import product
+from typing import List
+
+
+def check_sum_of_four(a: List[int], b: List[int], c: List[int], d: List[int]) -> int:
+    prod = list(product(a, b, c, d))
+    sums = []
+    for i in range(len(prod)):
+        sums.append(sum(prod[i]))
+    return sums.count(0)
