+++ homework1/task05/__init__.py
@@ -1,0 +1,27 @@
+"""
+Given a list of integers numbers "nums".
+
+You need to find a sub-array with length less equal to "k", with maximal sum.
+
+The written function should return the sum of this sub-array.
+
+Examples:
+    nums = [1, 3, -1, -3, 5, 3, 6, 7], k = 3
+    result = 16
+"""
+from typing import List
+
+
+def find_maximal_subarray_sum(nums: List[int], k: int) -> int:
+    if k > len(nums):
+        print("Given subarray length > array length")
+        pass
+    else:
+        sums = []
+        k_sum = 0
+        for i in range(len(nums) - (k - 1)):
+            for j in range(k):
+                k_sum += nums[i + j]
+            sums.append(k_sum)
+            k_sum = 0
+        return max(sums)
