+++ homework1/task05/tests/test_find_maximal_subarray_sum.py
@@ -1,4 +1,4 @@
-from task05 import find_maximal_subarray_sum
+from homework1.task05 import find_maximal_subarray_sum
 
 
 def test_maximal_subarray_sum_happy_path():
